Rename to implemented_remotes.py
as it now stores remote class definitions, not api responses
